# openapi.yaml
openapi: 3.0.0
info:
  title: 'LookDesk API'
  version: 1.0.0
servers:
  -
    url: 'https://api.lookdesk.ai'
    description: 'Production API'
paths:
...
  /api/billing/subscription:
    get:
      tags:
        - Billing
      summary: 'Get current user subscription'
      operationId: de186396f33048dbf960fc5d6900c65a
      responses:
        '200':
          description: 'Subscription data'
        '401':
          description: Unauthorized
      security:
        -
          sanctum: []
  /api/billing/subscribe:
    post:
      tags:
        - Billing
      summary: 'Create Stripe subscription'
      operationId: 5a8d8d87da2190bb603604ec497bf635
      requestBody:
        required: true
        content:
          application/json:
            schema: {  }
            example:
              plan_price_id: 1
              quantity: 1
      responses:
        '200':
          description: 'Checkout URL returned'
        '422':
          description: 'Validation error'
      security:
        -
          sanctum: []
  /api/billing/change-plan:
    put:
      tags:
        - Billing
      summary: 'Change subscription plan'
      operationId: e95b48f1cfe00b3a3cee470bc8f46cdd
      requestBody:
        required: true
        content:
          application/json:
            schema: {  }
            example:
              plan_price_id: 2
              quantity: 5
      responses:
        '200':
          description: 'Plan change requested'
        '400':
          description: 'No active subscription'
      security:
        -
          sanctum: []
  /api/billing/cancel:
    post:
      tags:
        - Billing
      summary: 'Cancel subscription'
      operationId: adb22729136ecd422b8450adf73a8331
      responses:
        '200':
          description: Canceled
        '400':
          description: 'No subscription'
      security:
        -
          sanctum: []
  /api/categories:
    get:
      tags:
        - Categories
      summary: 'Get all categories (with translations)'
      operationId: ac9c4d978ea4e5b72915ca9728e5b8b1
      parameters:
        -
          name: lang
          in: query
          schema:
            type: string
        -
          name: parent_id
          in: query
          schema:
            type: integer
            nullable: true
      responses:
        '200':
          description: 'Categories list'
      security:
        -
          sanctum: []
    post:
      tags:
        - Categories
      summary: 'Create new category'
      operationId: cca49a87d9b0ebb9a98ad7c154b71737
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - translations
              properties:
                parent_id:
                  type: integer
                  nullable: true
                sort_order:
                  type: integer
                translations:
                  type: array
                  items: { required: [lang, title], properties: { lang: { type: string }, title: { type: string }, description: { type: string, nullable: true } }, type: object }
              type: object
      responses:
        '201':
          description: Created
      security:
        -
          sanctum: []
  '/api/categories/{id}':
    get:
      tags:
        - Categories
      summary: 'Get category by ID'
      operationId: 09767bb2e47dde77dd1e2591428b0ae1
      parameters:
        -
          name: id
          in: path
          schema:
            type: integer
        -
          name: lang
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 'Category found'
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    put:
      tags:
        - Categories
      summary: 'Update category'
      operationId: 7dcf71887e585de7f881e9eb98d86126
      requestBody:
        content:
          application/json:
            schema:
              properties:
                parent_id:
                  type: integer
                  nullable: true
                sort_order:
                  type: integer
                translations:
                  type: array
                  items: { properties: { lang: { type: string }, title: { type: string }, description: { type: string, nullable: true } }, type: object }
              type: object
      responses:
        '200':
          description: Updated
      security:
        -
          sanctum: []
    delete:
      tags:
        - Categories
      summary: 'Delete category'
      operationId: 9b61b65823f4a21f182506b34ddf805c
      responses:
        '200':
          description: Deleted
      security:
        -
          sanctum: []
...
components:
...
  securitySchemes:
    sanctum:
      type: http
      bearerFormat: JWT
      scheme: bearer
...